Rename the similarity module to similarity_model
While at it, rename SimilarityModel::query to similarity.
