+++ src/inspector_window.rs
@@ -7,7 +7,7 @@ use gtk::{ApplicationWindow, Builder, Label, MenuItem, Notebook, Widget};
 use crate::analogy_widget::AnalogyWidget;
 use crate::metadata_dialog::MetadataDialog;
 use crate::metadata_model::MetadataModel;
-use crate::similarity::SimilarityModel;
+use crate::similarity_model::SimilarityModel;
 use crate::SimilarityWidget;
 
 pub struct InspectorWindow {
